move source nodes to shared events code
