+++ templates/components/multiagent/python/app/workflows/events.py
@@ -1,7 +1,10 @@
 from enum import Enum
-from typing import Optional
+from typing import List, Optional
 
+from llama_index.core.schema import NodeWithScore
 from llama_index.core.workflow import Event
+
+from app.api.routers.models import SourceNodes
 
 
 class AgentRunEventType(Enum):
@@ -23,5 +26,20 @@ class AgentRunEvent(Event):
                 "type": self.event_type.value,
                 "text": self.msg,
                 "data": self.data,
+            },
+        }
+
+
+class SourceNodesEvent(Event):
+    nodes: List[NodeWithScore]
+
+    def to_response(self):
+        return {
+            "type": "sources",
+            "data": {
+                "nodes": [
+                    SourceNodes.from_source_node(node).model_dump()
+                    for node in self.nodes
+                ]
             },
         }
